refact: Refactorizando NewPasswordController
